+++ packages/environment-check/test/checks/environment.test.ts
@@ -5,7 +5,7 @@ import { checkEndpoint } from '../../src/checks/endpoint';
 import { checkStoredSystems } from '../../src/checks/stored-system';
 import { DevelopmentEnvironment, Severity } from '../../src/types';
 import { isAppStudio } from '@sap-ux/btp-utils';
-import { join } from 'path';
+import { join } from 'node:path';
 import * as install from '../../src/checks/get-installed';
 
 jest.mock('axios');